fix: handle multiple type of file uploading
